Created PoopRecipes.js unification
There's now a PoopRecipes object (with prototypal inheritance and
everything hurray!) that abstracts all the work, so to add a site a
query selector is all that's needed.

The PoopRecipes object manages hashing of ingredient strings, so
although it seems random each ingredient will always give the same poop
substance. Poopification.